20230711 Hoàng . Get địa chỉ + get Fee from Giao hàng tiết kiệm

+++ src/main/java/com/datn/dongho5s/Entity/HoaDonChiTiet.java
@@ -42,7 +42,7 @@ public class HoaDonChiTiet {
     private Integer soLuong;
 
     @Column(name = "gia_ban")
-    private Double giaBan;
+    private Integer giaBan;
 
 
 }
